Develop the interface of the Cache further
We now also provide transaction objects for the Cache itself, like
the LMDB Safe library does.

This allows users of the Cache to execute multiple things in a
consistent fashion. A prime example is the name() method which
will often require the user to look up the parent ino before using
it.