Refactor Inode structs to make them POD
This allows us to directly reinterpret them from LMDB values.

Note that LMDB does not guarantee us any alignment beyond two, but
on amd64, it should be fine nevertheless (because the alignment
happens to work if keys have the correct size).

We might want to add assertions for that, or make parse handle
those cases by creating a copy anyways.