Fully sync on opendir(), including removal of stale entries
The previous code would only ever add new entries, never remove
them. This commit fixes that by introducing a rewrite mode for
directories (in the cache) allowing to indicate to the cache that
a directory is about to be completely rewritten, and all entries
which are not touched during the rewrite should be purged at the
end of the rewrite.