Ensure that st_size is correctly set for symlinks
The OS (and by extension, userland) expects to be able to read the
length of a link using the st_size field returned by lstat, so we
have to ensure that it is always consistent.

Note tha this implementation does not ensure this yet, as we do
not update/modify the link destination when a new set of attributes
is written for the inode. We would in theory have to e.g. truncate
the link if the new size written from the attributes is smaller,
or we would have to never overwrite the size of a link inode.